Split up the LeagueSeasonFilterBuilder and LeagueSeasonFilter
