+++ FootballHistory/Server/Builders/Models/LeagueSeasonFilter.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace FootballHistory.Server.Builders.Models
+{
+    public class LeagueSeasonFilter
+    {
+        public List<string> AllSeasons;
+        public List<Tier> AllTiers;
+    }
+}
